Fix CustomJWTAuthentication circular import issue
